Fixed several quality issues.

+++ src/test/java/org/javamoney/moneta/spi/DefaultNumberValueTest.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (c) 2012, 2015, Credit Suisse (Anatole Tresch), Werner Keil and others by the @author tag.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
@@ -151,7 +151,7 @@ public class DefaultNumberValueTest {
 
 		assertEquals(Long.valueOf(21L), numberValue.numberValue(Long.class));
 		assertEquals(Integer.valueOf(21),numberValue.numberValue(Integer.class));
-		assertEquals(Float.valueOf(21),numberValue.numberValue(Float.class));
+		assertEquals(21f,numberValue.numberValue(Float.class));
 		assertEquals(Byte.valueOf((byte)21),numberValue.numberValue(Byte.class));
 		assertEquals(Short.valueOf((short)21),numberValue.numberValue(Short.class));
 		assertEquals(new AtomicLong(21L).get(),numberValue.numberValue(AtomicLong.class).get());
@@ -167,7 +167,7 @@ public class DefaultNumberValueTest {
 
 		assertEquals(Long.valueOf(21L), numberValue.numberValueExact(Long.class));
 		assertEquals(Integer.valueOf(21),numberValue.numberValueExact(Integer.class));
-		assertEquals(Float.valueOf(21),numberValue.numberValueExact(Float.class));
+		assertEquals(21f,numberValue.numberValueExact(Float.class));
 		assertEquals(Byte.valueOf((byte)21),numberValue.numberValueExact(Byte.class));
 		assertEquals(Short.valueOf((short)21),numberValue.numberValueExact(Short.class));
 		assertEquals(new AtomicLong(21L).get(),numberValue.numberValueExact(AtomicLong.class).get());
